refactor(user): remove UserService interface
