--- tracker-api/src/main/java/com/sloyardms/trackerapi/user/UserServiceImpl.java
@@ -1,72 +1,0 @@
-package com.sloyardms.trackerapi.user;
-
-import com.sloyardms.trackerapi.user.dto.UserCreateDto;
-import com.sloyardms.trackerapi.user.dto.UserDto;
-import com.sloyardms.trackerapi.user.dto.UserUpdateDto;
-import com.sloyardms.trackerapi.user.entity.User;
-import com.sloyardms.trackerapi.common.exception.ResourceDuplicatedException;
-import com.sloyardms.trackerapi.common.exception.ResourceNotFoundException;
-import com.sloyardms.trackerapi.user.mapper.UserMapper;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.UUID;
-
-@Service
-public class UserServiceImpl implements UserService {
-
-    private final UserRepository userRepository;
-    private final UserMapper userMapper;
-
-    public UserServiceImpl(UserRepository userRepository, UserMapper userMapper) {
-        this.userRepository = userRepository;
-        this.userMapper = userMapper;
-    }
-
-    @Override
-    @Transactional(rollbackFor = Exception.class)
-    public UserDto create(UserCreateDto userDto) {
-        if(userRepository.existsById(userDto.getUuid())){
-            throw new ResourceDuplicatedException("UUID already exists");
-        }
-
-        User user = userMapper.toEntity(userDto);
-
-        try {
-            User savedUser = userRepository.saveAndFlush(user);
-            return userMapper.toDto(savedUser);
-        }catch (Exception e){
-            throw new ResourceDuplicatedException("Username already exists", e);
-        }
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public UserDto findByUuid(UUID uuid) {
-        User userDb = userRepository.findById(uuid).orElseThrow(() -> new ResourceNotFoundException("User with UUID " + uuid + " not found"));
-        return userMapper.toDto(userDb);
-    }
-
-    @Override
-    @Transactional(rollbackFor = Exception.class)
-    public UserDto update(UUID uuid, UserUpdateDto userDto) {
-        User userDb = userRepository.findById(uuid).orElseThrow(() -> new ResourceNotFoundException("User with UUID " + uuid + " not found"));
-
-        userMapper.updateFromDto(userDto, userDb);
-
-        try {
-            User updatedUser = userRepository.saveAndFlush(userDb);
-            return userMapper.toDto(updatedUser);
-        }catch (Exception e){
-            throw new ResourceDuplicatedException("Username already exists", e);
-        }
-    }
-
-    @Override
-    @Transactional(rollbackFor = Exception.class)
-    public void delete(UUID uuid) {
-        userRepository.findById(uuid).orElseThrow(() -> new ResourceNotFoundException("User with UUID " + uuid + " not found"));
-        userRepository.deleteById(uuid);
-    }
-
-}
